+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Changes the parsed tag name of the default XML namespace from `$foo` to `:foo` to better match other unknown namespaces.
 - Adds `--namespace` argument to the CLI.
+- Fixes overriding of tag attribute when its nested tag has the same name.
 
 ## 0.1.338 (Feb 6, 2017)
 - Fixes `hugml --version`.