+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 1.0.0 (Sep 2, 2017)
 - Changes the parsed tag name of the default XML namespace from `$foo` to `:foo` to better match other unknown namespaces.
 - Adds `--namespace` argument to the CLI.
 - Fixes overriding of tag attribute when its nested tag has the same name.